backend/requirements.txt
# ==============================
# Core Framework
# ==============================
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.7.4
pydantic-settings==2.6.0

# ==============================
# Database Drivers
# ==============================
pymongo==4.6.1
neo4j==5.16.0
mysql-connector-python==8.3.0
sqlalchemy==2.0.25
greenlet==3.0.3

# ==============================
# SQL Parsing
# ==============================
sqlparse==0.4.4
sqlglot==20.11.0

# ==============================
# AI / ML APIs
# ==============================
langgraph==0.2.45
langchain-core==0.3.21
openai>=1.0.0
# langchain-community==0.3.8  # Commented out due to numpy compilation issues

# ==============================
# Utilities
# ==============================
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.26.0
pyyaml==6.0.1
email-validator==2.1.0.post1
orjson==3.9.15
tenacity==8.2.3